Now downloads dives from computer and creates UDDF
The UDDF generation isn't perfect yet. In particular it doesn't
properly identify when a dive has ended and it groups all the dives
together into a single repetition group. Also, when importing
UDDF dives into MacDave it appears that it loses the information
about the date and time of the dive.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>