Added vendor and event output, switch to GDateTime
Vendor and Event tags are now output with the UDDF file even though
they're not standard. It's helpful for me to understand what is going
on in my dives if I can see this information.

Internally the date format was switched over to a GDateTime format
for logging the start of dives. This brings in support for timezones.
By default a dive is done is local time, although it is possible to use
dif_dive_set_datetime_utc to set the time in UTC. Most dive computers,
however are set to the local time of the diver.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>