Implemented initial pressure fix algorithm
This resides in the new dif/algos.c file and can be accessed using
the following two methods:

dif_alg_dc_initial_pressure_fix (for an entire dif_dive_collection_t)
dif_alg_dive_initial_pressure_fix (for just a dif_dive_t)

The algorithm is fairly naive. It scans through the samples looking
for the first time that there is a pressure with a value of greater
than 1.0 bar. It saves that value then it goes back through the
samples and sets the pressure of those initial samples that were
less than 1.0 bar to the new pressure.

Closes #6
Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>