More robust support for gasmixes
In particular this created the enum `dif_gasmix_type_t` and also
the function `dif_gasmix_type`. Right now it works for air and
most flavors of nitrox.

This is needed because the gas definitions in UDDF are stupidly
at the head of the dive file rather than associated with each dive.

Addresses part of #2

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>