Support for gas mixes from dives
This was a little tricky as the gas mixes are always done via reference
from each dive. I guess this makes sense overall, but it's done via a
<link> element which can also be used to reference tanks. Ughh. What
a mess.

Anyway, the <tankdata> elements are very rudimentary. I don't bother
with any of the <tankbeginpressure> or <tankendpressure> because
those values can be easily derived.

The other tricky thing is that Uwatec Galileo computers seem to
always report three tanks, with tanks 2 and 3 being 100% N2 if they're
not used. Thus I had to create a function called `dif_gasmix_is_valid`
to check for these cases and tell it not to save those gasmixes.
However, other gasmixes that are 100% of a gas (e.g. 100% O2 or 100%
Ar) are valid.

Closes #2

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>