Cleaned up XML output
The code almost validates, however the UDDF spec requires that all files
have a `<diver>` element. This isn't something that we can easily do
because we don't actually know anything about the diver.

As a result of this, I've created the `<informationafterdive>` element
and have done more work to ensure that the files validate against the
current version of the schema (see issue #13).

There still is more work to be done on `<informationafterdive>` until
I can close issue #12. Particularly `<pressuredrop>` and `<averagedepth>`
need to be calculated.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>