chore(bump): upgrade libdivecomputer to 0.9.0
This bumps the required version of libdivecomputer from 0.3.0, which was
the version in 2012, to 0.9.0, which is the current version for 2025. As
a result, it brings in a ton of new dive computers, while still
supporting many of the older ones too.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>